+++ tidy3d/components/structure.py
@@ -678,7 +678,8 @@ class MeshOverrideStructure(AbstractStructure):
         title="Grid Size Choice In Structure Overlapping Region",
         description="In structure intersection region, grid size is decided by the latter added "
         "structure in the structure list when ``shadow=True``; or the structure of smaller grid size "
-        "when ``shadow=False``.",
+        "when ``shadow=False``. If ``shadow=False``, and the structure doesn't refine the mesh, grid snapping to "
+        "the bounding box of the structure is disabled.",
     )
 
     @pydantic.validator("geometry")